TransformUtils: Fix reverse deletion loop in SinkThenElseCodeToEnd
The previous logic was effectively:

    for (auto RI = BB.rbegin(); RI != BB.rend();)
      RI->eraseFromParent();

With std::reverse_iterator, this worked.  RI had a handle to
RI->getNext(), and called getPrev() on dereference... deleting the node
*RI would magically "iterate" RI and change the result of BB.rend().

Now that ilist uses a proper reverse_iterator (r280032), this magic is
broken.  This commit fixes it:

    for (auto RI = BB.rbegin(), RE = BB.rend(); RI != RE;)
      RI++->eraseFromParent();

The changes (required by ilist changes):
- Incrementing is now explicit.
- rend() is now stable.

The logic is actually pretty hard to tease out of the code (the loop
could be much, much simpler), but upstream has moved on so there's no
point in cleaning it up.  After we branched for swift-3.1-branch, James
Malloy completely rewrote this (and unfortunately, no, there wasn't a
targeted commit to cherry-pick).